iio: frequency : adf4360: Move definitions to implementation file since they are specific and not common.
